Desktop/Fall 2021/python/udacity_project2/requirements.txt
black==21.9b0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
flask==2.0.1
requests==2.26.0
typing-extensions==3.10.0.2
python_docx==0.8.10
pdftotext==2.2.0
pep257==0.7.0